requirements.txt
# Add your dependencies here
fastapi
uvicorn[standard]
pydantic>=2
reportlab
arabic-reshaper
python-bidi
Pillow
streamlit
requests
python-multipart
